+++ nodes/KlickTipp/actions/opt-in-process/get-redirect.operation.ts
@@ -16,12 +16,11 @@ export const properties: INodeProperties[] = [
 			'Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code-examples/expressions/">expression</a>',
 	},
 	{
-		displayName: 'Email',
+		displayName: 'Email Address',
 		name: 'email',
 		type: 'string',
 		placeholder: 'Enter email address (required)',
 		default: '',
-		description: 'Email address',
 	},
 ];
 
@@ -39,11 +38,11 @@ export async function execute(this: IExecuteFunctions, index: number) {
 	const email = this.getNodeParameter('email', index) as string;
 
 	if (!listId) {
-		return handleError.call(this, 'The opt-in process ID is required.');
+		return handleError.call(this, 'Opt-in process ID is missing');
 	}
 
 	if (!email) {
-		return handleError.call(this, 'The email address is required.');
+		return handleError.call(this, 'Email is missing');
 	}
 
 	const body: IDataObject = {