+++ nodes/KlickTipp/actions/subscriber/index.ts
@@ -6,29 +6,10 @@ import * as index from './index.operation';
 import * as del from './delete.operation';
 import * as get from './get.operation';
 import * as search from './search.operation';
-import * as signIn from './sign-in.operation';
-import * as signOff from './sign-off.operation';
-import * as signOut from './sign-out.operation';
-import * as tag from './tag.operation';
 import * as tagged from './tagged.operation';
-import * as untag from './untag.operation';
 import * as update from './update.operation';
 
-export {
-	subscribe,
-	unsubscribe,
-	index,
-	del as delete,
-	search,
-	get,
-	signIn,
-	signOff,
-	signOut,
-	tag,
-	tagged,
-	untag,
-	update,
-};
+export { subscribe, unsubscribe, index, del as delete, search, get, tagged, update };
 
 export const description: INodeProperties[] = [
 	{
@@ -43,88 +24,53 @@ export const description: INodeProperties[] = [
 		},
 		options: [
 			{
-				name: 'Subscriber Delete',
-				value: 'delete',
-				description: 'Deletes a subscriber',
-				action: 'Subscriber delete',
-			},
-			{
-				name: 'Subscriber Get',
-				value: 'get',
-				description:
-					'Returns all available information about a subscriber. For example, you can use it to check when a contact received, opened, or clicked in a specific email.',
-				action: 'Subscriber get',
-			},
-			{
-				name: 'Subscriber Index',
-				value: 'index',
-				description:
-					'Returns an associative array in the format <code>{ subscriber_id: ID }</code>, where each key corresponds to a unique subscriber ID. Use the <code>Subscriber Get</code> node to retrieve detailed information about a subscriber.',
-				action: 'Subscriber index',
-			},
-			{
-				name: 'Subscriber Search',
-				value: 'search',
-				description:
-					'Returns the subscriber ID for an email address. Use the <code>Subscriber Get</code> node to retrieve detailed information about this subscriber.',
-				action: 'Subscriber search',
-			},
-			{
-				name: 'Subscriber Signin',
-				value: 'signIn',
-				description:
-					'Adds or updates a subscriber using an API key. The subscriber can be identified by their email address or SMS number, with optional fields.',
-				action: 'Subscriber signin',
-			},
-			{
-				name: 'Subscriber Signoff',
-				value: 'signOff',
-				description: 'Unsubscribes a subscriber using an API key',
-				action: 'Subscriber signoff',
-			},
-			{
-				name: 'Subscriber Signout',
-				value: 'signOut',
-				description: 'Removes the manual tag or SmartLink defined in the API key from a subscriber',
-				action: 'Subscriber signout',
-			},
-			{
-				name: 'Subscriber Subscribe',
+				name: 'Add or Update Contact',
 				value: 'subscribe',
 				description:
-					'Adds or updates a subscriber. The subscriber can be identified by their email address or SMS number, with an optional opt-in process, tag, and fields.',
-				action: 'Subscriber subscribe',
+					'Adds a new contact. If a contact with the same email already exists, it will be updated.',
+				action: 'Add or update contact',
 			},
 			{
-				name: 'Subscriber Tag',
-				value: 'tag',
-				description: 'Tags a subscriber with a manual tag or SmartLink',
-				action: 'Subscriber tag',
+				name: 'Delete Contact',
+				value: 'delete',
+				description: 'Deletes a contact',
+				action: 'Delete contact',
 			},
 			{
-				name: 'Subscriber Tagged',
+				name: 'Get Contact',
+				value: 'get',
+				description: 'Returns the complete data of a contact',
+				action: 'Get contact',
+			},
+			{
+				name: 'Get Contact ID',
+				value: 'search',
+				description: 'Returns the contact ID for an email address',
+				action: 'Get contact id',
+			},
+			{
+				name: 'List Contacts',
+				value: 'index',
+				description: 'Lists the IDs of all active contacts',
+				action: 'List contacts',
+			},
+			{
+				name: 'List Tagged Contacts',
 				value: 'tagged',
-				description:
-					'Returns an associative array in the format <code>{ subscriber_id: subscription_timestamp }</code>, where each key represents the unique ID of a tagged subscriber, and the value is the timestamp when the subscription occurred. Use the <code>Subscriber Get</code> node to retrieve detailed information about a subscriber.',
-				action: 'Subscriber tagged',
+				description: 'Lists the IDs and tagging timestamps of all tagged contacts',
+				action: 'List tagged contacts',
 			},
 			{
-				name: 'Subscriber Unsubscribe',
+				name: 'Unsubscribe Contact',
 				value: 'unsubscribe',
-				description: 'Unsubscribes a subscriber',
-				action: 'Subscriber unsubscribe',
+				description: 'Unsubscribes a contact, preventing further communication',
+				action: 'Unsubscribe contact',
 			},
 			{
-				name: 'Subscriber Untag',
-				value: 'untag',
-				description: 'Removes a manual tag or SmartLink from a subscriber',
-				action: 'Subscriber untag',
-			},
-			{
-				name: 'Subscriber Update',
+				name: 'Update Contact',
 				value: 'update',
-				description: 'Updates the data fields, email address, or SMS number of a subscriber',
-				action: 'Subscriber update',
+				description: 'Updates a contact',
+				action: 'Update contact',
 			},
 		],
 		default: 'index',
@@ -135,11 +81,6 @@ export const description: INodeProperties[] = [
 	...del.description,
 	...get.description,
 	...search.description,
-	...signIn.description,
-	...signOff.description,
-	...signOut.description,
-	...tag.description,
 	...tagged.description,
-	...untag.description,
 	...update.description,
 ];